Significantly improve schema validation and autocomplete by introducing InputValue and friends.
This cleans up the types used for schema generation to only things that can legally be represented in JSON.

The `instanceOf(Embedding)` type was messing up the seed packet schema with an 'any-like' pattern, which basically disabled schema validation in sub-seeds. For some reason `strictUnions` in
zod-to-json-schema wasn't working (causing an error when it was stripped).

This is a bit hacky, but still should roughly work.

Part of #7.