Add an ability for a seed packet to define an environment overlay, that is overlaid on top of the default environment for each seed in the packet before being executed as a top-level seed.
Note that there is a non-intuitive behavior, documented in a TODO in README, where seeds executed as
sub-seeds of parents in other packets will not have the base environment they expect, since currently the
enviroment for a sub-seed is entirely defined by the calller seed, which might come from a different
context. That makes this feature less useful than it might otherwise be, because seeds that are
entrypoints can't rely on having baseline variables set.

Part of #38.