GNARLY COMMIT. When nested seed data includes sub-seeds, automatically inject necessary `array` and `object` seeds during seed packet expansion.
The upshot is that you can now skip including manual `object` or `array` seed types in seed defintions and they will be added for you.

One gotcha is that sub-objects cannot contain the properties `seed` or `type`, otherwise they will be interpreted as a sub_seed. (This was technically always true though).

This is a gnarly commit with a terrifyingly complex expandSeedComputedObjects method. The only reason I have any confidence whatsoever that it works is that tests pass.

Part of #42.