Make `mock` environment variable a `protected` environment var.
This means you may only retrieve it from `getKnownProtectedKey`. Also, it means that environment.clone()
(which is used by let, let-multi, and packet.environment) will only allow setting to true, never to
false.

This behavior is useful for sensitive keys where a calling seed wants to force behavior of a sub-graph.

Part of #53.