Store stores and memory ids in filesystem profile.
This is necessary in order to create an `enumerate` seed_type soon.

We run every memoryID and storeID through `safeName`, which is a one-way mangle of the name, so to enumerate ids we need to store them whwn we save.

Note that this will make old profiles before this change fail to load and validate. I'm not bothering to update the profile metadata version number and do upgrading machinery for this small change.
If you have a profile, delete it before running.

Part of #55.