Make it so common properties of seeds show up last in JSON autocomplete.
Those properties (e.g. id, private, comment, description) are true for every seed. Before they would show up first, but they are the least interesting to remind you which properties are the most
important for a given seed type.

Now autocomplete will be more helpful with the seed-type-specific properties first.

Note that at this moment schema autocompleting isn't working for me, but eyeballing the schema it looks like it should work.

Part of #7.