Make it so selectGarden only loads the currentPacket.
Before, if there was any validation error anywhere in the packets, it wouldn't allow you to run any seed.

Now, as long as there aren't validation errors in the current packet, it's fine.

This means that local-to-local packet references temporarily won't work, but soon we'll add our own Profile that knows how to fetch local packets by name from state.

Part of #6.