Add a changeSeedType meta function. Also add SeedDataIsh and argumentsAndOptions.
changeSeedType changes a SeedData to have a different type, and adds missing required properties and options.

SeedDataIsh is a new type that is required because the result of changeSeedType might return a technically invalid seed data.

argumentsAndOptions is a very straightforward convenience helper for iterating through arguments and options of SeedShape.

Part of #6.