Fix a bug where pickPacketAndSeed didn't work to select when nothing currently set.
it was using the result of firstNonEmptyPacketName and merging it in, but the property names were the wrong names!

Typescript didn't complain beacuse the type of result was set explicitly, and so as far as typescript is concerned those new names would just be ignored because they were beyond the type
definition.

Part of #6.