DATA MODEL CHANGE. Add a collapsed property to PacketWrapper.
This will be used to store whether the zippy is collapsed or open in seed-list.

Note again that this breaks currently persited seed packets in localStorage since they won't have a
required collapsed parameter, but since I'm the only one using it right now, I won't bother writing
a migration script.

Part of #6.