Short circuit updateHash earlier.
I think this behavior is justified. Before, when the hash was updated manually, updateHash would be called, which would then call window.location.hash... which would call updateHash a second time
inside the first.

This SHOULD reduce extra UPDATE_HASH events. I think it works fine in all cases, because even if it comes from the URL, if it's already what's in state then it doesn't need to be updated.

Part of #6.