Make it so seed-editor disables the ability to remove a type.
After the last commit, the 'delete' button next to type looked weird, like it went on the wrong side
of the select. That's because type doesn't have a typeChange, and also has a select for seed type.

This makes it look cleaner. If you want to change a seed to a generic object, you can switch the
type of the object to obejct.

Part of #6.