Change the way we decide which dialog gets keyboard events.
Before we assumed the order of when it was opened determined who won. But really it's the order of them in
the DOM, with the last one being the most likely to be on top.

Note that this is also not strictly true; it assumes that dialogs are all stamped out within the same
parent, in order, with no z-index trickery or antyhing.

Part of #6.