Big commit. Create TypeShape and use that for PropertyShape. Add a concept of an 'unknown'/any TypeShape.
Currently TypeShape is basically just an indirected wrapper for PropertyType, but with the addition of an
'unknown' value, which means 'fine with any PropertyType'.

This changes a bit of the logic for when and how to render drop downs, which I believe is working.

This work will allow in the near future specifying sub-object and sub-array types instead of just "it's an
array :shrug:"

Part of #6.